+++ canoe/provider/src/lib.rs
@@ -1,8 +1,5 @@
-pub mod canoe_input;
-pub use canoe_input::CanoeInput;
-
 pub mod canoe_provider;
-pub use canoe_provider::{CanoeNoOpProvider, CanoeProvider};
+pub use canoe_provider::{CanoeInput, CanoeNoOpProvider, CanoeProvider};
 
 pub mod verifier_caller;
 pub use verifier_caller::CertVerifierCall;
