+++ canoe/sp1-cc/verifier/src/lib.rs
@@ -27,7 +27,7 @@ use tracing::{info, warn};
 /// ```
 /// The v_key will be printed in the terminal.
 pub const V_KEY: [u32; 8] = [
-    1184335638, 1135783716, 1423207388, 911618791, 1295955408, 1041966695, 440842582, 38998658,
+    155840827, 1054807976, 1117502010, 411643090, 1737149458, 1390688513, 1427847064, 415719986,
 ];
 
 #[derive(Clone)]
